requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==5.0.9
h5py==3.2.1
Hydra==2.5
hydra-core==1.0.6
idna==2.10
imbalanced-learn==0.8.0
importlib-resources==5.1.4
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.4.2
numba==0.53.1
numpy==1.20.3
omegaconf==2.0.6
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pooch==1.3.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
torch==1.8.1
torchaudio==0.8.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
